requirements.txt
pyttsx3
requests
psutil
speechrecognition
selenium
python-dotenv
webdriver-manager
google-generativeai
